\name{addRecipient}
\alias{addRecipient}
\title{Adds a recipient to the given panel.}
\usage{
  addRecipient(username, password, libraryId, panelId,
    firstName = NULL, lastName = NULL, email = NULL,
    externalDataRef = NULL, embeddedData = NULL)
}
\arguments{
  \item{username}{the Qualtrics username.}

  \item{password}{the Qualtrics password.}

  \item{libraryId}{the Qualtrics library id.}

  \item{panelId}{the Qualtrics panel id.}

  \item{firstName}{the recipient's first name.}

  \item{lastName}{the recipient's last name.}

  \item{email}{the recipient's email address.}

  \item{externalDataRef}{Qualtrics extrenral data ref.}

  \item{embeddedData}{embedded data for the recipient.}
}
\description{
  Adds a recipient to the given panel.
}

