% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_mailinglists.R
\name{all_mailinglists}
\alias{all_mailinglists}
\title{Retrieve a data frame of all mailing lists from Qualtrics}
\usage{
all_mailinglists()
}
\description{
Retrieve a data frame of all mailing lists from Qualtrics
}
\examples{
\dontrun{
mailinglists <- all_mailinglists()
}
}
