% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{append_root_url}
\alias{append_root_url}
\title{Append proper end points to create root URL}
\usage{
append_root_url(base_url, type = c("responseexports", "surveys"))
}
\arguments{
\item{base_url}{Base URL for your institution (see
\url{https://api.qualtrics.com/docs/root-url}}

\item{type}{Either `"responseexports"` or `"surveys"`}
}
\value{
Root URL, plus appended end point
}
\description{
Append proper end points to create root URL
}
