% Generated by roxygen2 (4.0.1): do not edit by hand
\name{auth}
\alias{auth}
\title{authenticate}
\usage{
auth(url = "https://survey.qualtrics.com/WRAPI/ControlPanel/api.php",
  useremail, brand = "", token)
}
\description{
This function creates a list with authentication details to access the
Qualtrics API
}

