% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{cached_file_download}
\alias{cached_file_download}
\title{Cached wrapper around export_responses_filedownload()}
\usage{
cached_file_download(surveyID, fileID, tmp_dir)
}
\arguments{
\item{surveyID, fileID, tmp_dir}{Passed to \code{\link[=export_responses_filedownload]{export_responses_filedownload()}}.}
}
\description{
Cached wrapper around export_responses_filedownload()
}
\keyword{internal}
