% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_existing_download}
\alias{check_existing_download}
\title{Check if downloaded file already exists}
\usage{
check_existing_download(file_location, surveyID, verbose = TRUE)
}
\arguments{
\item{file_location}{(potential) path to previous download}

\item{surveyID}{Qualtrics survey ID}

\item{verbose}{whether to report if match is found}
}
\description{
Check if downloaded file already exists
}
\keyword{internal}
