% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_for_warnings}
\alias{check_for_warnings}
\title{Check if httr GET result contains a warning}
\usage{
check_for_warnings(resp)
}
\arguments{
\item{resp}{object returned by \code{\link[=qualtrics_response_codes]{qualtrics_response_codes()}}}
}
\description{
Check if httr GET result contains a warning
}
\keyword{internal}
