% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_params}
\alias{check_params}
\title{Check if parameters passed to functions are correct}
\usage{
check_params(...)
}
\arguments{
\item{...}{options passed to function}
}
\description{
Check if parameters passed to functions are correct
}
