% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{checkarg_col_types}
\alias{checkarg_col_types}
\title{col_types must be a col_spec object from readr}
\usage{
checkarg_col_types(col_types)
}
\description{
col_types must be a col_spec object from readr
}
\keyword{internal}
