% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{checkarg_elements}
\alias{checkarg_elements}
\title{Check if elements given in fetch_description are properly specified}
\usage{
checkarg_elements(elements)
}
\description{
Check if elements given in fetch_description are properly specified
}
\keyword{internal}
