% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{checkarg_fetch_id_data}
\alias{checkarg_fetch_id_data}
\title{Check if data for fetch_id() is correct}
\usage{
checkarg_fetch_id_data(.data)
}
\description{
Check if data for fetch_id() is correct
}
\keyword{internal}
