% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{checkarg_file_name}
\alias{checkarg_file_name}
\title{Check if survey file specified in file_name exists}
\usage{
checkarg_file_name(file_name)
}
\description{
Check if survey file specified in file_name exists
}
\keyword{internal}
