% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{checkarg_get}
\alias{checkarg_get}
\title{Check if elements given in metadata's 'get' are properly specified}
\usage{
checkarg_get(get)
}
\description{
Check if elements given in metadata's 'get' are properly specified
}
\keyword{internal}
