% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{checkarg_include_embedded}
\alias{checkarg_include_embedded}
\title{Check include_embedded and format for API call:}
\usage{
checkarg_include_embedded(include_embedded)
}
\description{
Check include_embedded and format for API call:
}
\keyword{internal}
