% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{checkarg_include_metadata}
\alias{checkarg_include_metadata}
\title{Check that include_metadata has the right elements & format for API call}
\usage{
checkarg_include_metadata(include_metadata)
}
\description{
Check that include_metadata has the right elements & format for API call
}
\keyword{internal}
