% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{checkarg_include_questions}
\alias{checkarg_include_questions}
\title{Check that include_questions uses the QID format, and format for API call:}
\usage{
checkarg_include_questions(include_questions)
}
\description{
Check that include_questions uses the QID format, and format for API call:
}
\keyword{internal}
