% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{checkarg_ischaracter}
\alias{checkarg_ischaracter}
\title{Is character vector with no missing values:}
\usage{
checkarg_ischaracter(arg, null_okay = TRUE)
}
\arguments{
\item{arg}{argument to be checked}

\item{null_okay}{Boolean. whether argument can be NULL}
}
\description{
Is character vector with no missing values:
}
\keyword{internal}
