% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{checkarg_isintegerish}
\alias{checkarg_isintegerish}
\title{Is integerish (length-1 numeric w/o any fractional component)}
\usage{
checkarg_isintegerish(arg, null_okay = TRUE)
}
\description{
Is integerish (length-1 numeric w/o any fractional component)
}
\keyword{internal}
