% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{checkarg_isnumeric}
\alias{checkarg_isnumeric}
\title{Is a length-1 non-negative numeric scalar}
\usage{
checkarg_isnumeric(arg, null_okay = TRUE)
}
\description{
Is a length-1 non-negative numeric scalar
}
\keyword{internal}
