% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{checkarg_survey_active}
\alias{checkarg_survey_active}
\title{Check if active from update_survey_metadata() is boolean, then convert to active/inactive}
\usage{
checkarg_survey_active(active)
}
\description{
Check if active from update_survey_metadata() is boolean, then convert to active/inactive
}
\keyword{internal}
