% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{checkarg_time_zone}
\alias{checkarg_time_zone}
\title{Checks time zone, setting default if needed:}
\usage{
checkarg_time_zone(time_zone)
}
\description{
Checks time zone, setting default if needed:
}
\keyword{internal}
