% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{checkForWarnings}
\alias{checkForWarnings}
\title{Check if httr GET result contains a warning}
\usage{
checkForWarnings(resp)
}
\arguments{
\item{resp}{object returned by \code{\link{qualtrics_response_codes}}}
}
\description{
Check if httr GET result contains a warning
}
