% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{checkParams}
\alias{checkParams}
\title{Check if parameters passed to functions are correct}
\usage{
checkParams(...)
}
\arguments{
\item{...}{options passed to checkParams}
}
\description{
Check if parameters passed to functions are correct
}
