% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{constructHeader}
\alias{constructHeader}
\title{Construct a header to send to Qualtrics API}
\usage{
constructHeader(API_TOKEN)
}
\arguments{
\item{API_TOKEN}{API token. Available in your Qualtrics account (see: \url{https://api.qualtrics.com/docs/authentication})}
}
\description{
Construct a header to send to Qualtrics API
}
\seealso{
See \url{https://api.qualtrics.com/docs/root-url} for documentation on the Qualtrics API.
}
