% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_root_url}
\alias{create_root_url}
\title{Append to listed server to create root URL}
\usage{
create_root_url(base_url)
}
\arguments{
\item{base_url}{Base URL for your institution (see
\url{https://api.qualtrics.com/docs/root-url}}
}
\value{
Root URL
}
\description{
Append to listed server to create root URL
}
