% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panels.R
\name{createPanel}
\alias{createPanel}
\title{Creates a new panel.}
\usage{
createPanel(username, password, libraryId, name, category = NULL)
}
\arguments{
\item{username}{the Qualtrics username.}

\item{password}{the Qualtrics password.}

\item{libraryId}{the Qualtrics library id.}

\item{name}{the panel name.}

\item{category}{the panel's category (optional).}
}
\description{
Creates a new panel.
}

