% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{download_qualtrics_export}
\alias{download_qualtrics_export}
\title{Download response export}
\usage{
download_qualtrics_export(fetch_url, requestID, verbose = FALSE)
}
\arguments{
\item{fetch_url}{URL provided by Qualtrics API that shows the download percentage completeness}

\item{requestID}{ID}

\item{verbose}{See \code{\link{fetch_survey}}}
}
\description{
Download response export
}
\keyword{internal}
