% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{downloadQualtricsExport}
\alias{downloadQualtricsExport}
\title{Download response export}
\usage{
downloadQualtricsExport(check_url, verbose = FALSE)
}
\arguments{
\item{check_url}{URL provided by Qualtrics API that shows the download percentage completneness}

\item{verbose}{See \code{\link{getSurvey}}}
}
\description{
Download response export
}
