% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_survey.R
\name{export_responses_filedownload}
\alias{export_responses_filedownload}
\title{Downloads response data from location obtained from
fetch_survey_progress (extracting from .zip file)}
\usage{
export_responses_filedownload(surveyID, fileID, tmp_dir)
}
\arguments{
\item{surveyID}{survey ID}

\item{fileID}{file ID from fetch_survey_progress}

\item{tmp_dir}{temporary directory to use}
}
\description{
Downloads response data from location obtained from
fetch_survey_progress (extracting from .zip file)
}
\keyword{internal}
