% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_survey.R
\name{export_responses_init}
\alias{export_responses_init}
\title{Initiate a request to the export-responses API endpoint}
\usage{
export_responses_init(surveyID, body)
}
\arguments{
\item{surveyID}{ID of survey whose responses are being pulled}
}
\description{
Initiate a request to the export-responses API endpoint
}
\details{
If the request to the Qualtrics API made by this function fails, the request
will be retried. If you see these failures on a 500 error (such as a 504
error) be patient while the request is retried; it will typically succeed
on retrying. If you see other types of errors, retrying is unlikely to help.
}
\keyword{internal}
