% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_survey.R
\name{export_responses_progress}
\alias{export_responses_progress}
\title{Monitor progress from response request download, then obtain file download
location}
\usage{
export_responses_progress(surveyID, requestID, verbose = FALSE)
}
\arguments{
\item{surveyID}{ID of survey whose responses are being pulled}

\item{requestID}{exportProgressId from
https://api.qualtrics.com/37e6a66f74ab4-get-response-export-progress}

\item{verbose}{See \code{\link[=fetch_survey]{fetch_survey()}}}
}
\description{
Monitor progress from response request download, then obtain file download
location
}
\details{
If the request to the Qualtrics API made by this function fails, the request
will be retried. If you see these failures on a 500 error (such as a 504
error) be patient while the request is retried; it will typically succeed
on retrying. If you see other types of errors, retrying is unlikely to help.
}
\keyword{internal}
