% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_colmap.R
\name{extract_colmap}
\alias{extract_colmap}
\title{Helper function to extract the column map attached to a response data
download obtained from fetch_survey (using the default
\code{colmap_attrs = TRUE})}
\usage{
extract_colmap(respdata)
}
\arguments{
\item{respdata}{Response data including a column map dataframe as an attribute}
}
\description{
Helper function to extract the column map attached to a response data
download obtained from fetch_survey (using the default
\code{colmap_attrs = TRUE})
}
