% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_distributions.R
\name{fetch_distributions}
\alias{fetch_distributions}
\title{Download distribution data for a survey from Qualtrics}
\usage{
fetch_distributions(surveyID)
}
\arguments{
\item{surveyID}{String. Unique ID for the survey distribution data you want to download.}
}
\description{
Download distribution data for a survey from Qualtrics
}
\examples{
\dontrun{
# Register your Qualtrics credentials if you haven't already
qualtrics_api_credentials(
  api_key = "<YOUR-API-KEY>",
  base_url = "<YOUR-BASE-URL>"
)

surveys <- all_surveys()
distributions <- fetch_distributions(surveys$id[1])
}

}
