% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_id.R
\name{fetch_id}
\alias{fetch_id}
\title{Fetch a Qualtrics surveyID based on the survey name in the Qualtrics UI}
\usage{
fetch_id(x, survey_name)
}
\arguments{
\item{x}{Dataframe created by the function all_surveys().}

\item{survey_name}{Name of the survey as it appears in the Qualtrics UI. Must
be unique to be passed to fetch_id().}
}
\description{
Fetch a Qualtrics surveyID based on the survey name in the Qualtrics UI
}
\examples{
\dontrun{
# Register your Qualtrics credentials if you haven't already
qualtrics_api_credentials(
  api_key = "<YOUR-API-KEY>",
  base_url = "<YOUR-BASE-URL>"
)

# Retrieve a list of surveys
surveys <- all_surveys()

# Retrieve surveyID for a unique survey
my_id <- fetch_id(surveys, "Unique Survey Name")

}
}
