% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_mailinglist.R
\name{fetch_mailinglist}
\alias{fetch_mailinglist}
\title{Download a mailing list from Qualtrics}
\usage{
fetch_mailinglist(mailinglistID)
}
\arguments{
\item{mailinglistID}{String. Unique ID for the mailing list you want to
download. Returned as \code{mailingListId} by the
\link[=all_mailinglists]{all_mailinglists} function.}
}
\description{
Download a mailing list from Qualtrics
}
\details{
This function uses the Qualtrics XM Directory API. The directory it queries
is discovered automatically from your account. If your account has more than
one XM Directory, set the \code{QUALTRICS_DIRECTORY_ID} environment variable to
the directory ID you want to use; otherwise the first directory returned by
the API is used (with a warning).

If the request to the Qualtrics API made by this function fails, the request
will be retried. If you see these failures on a 500 error (such as a 504
error) be patient while the request is retried; it will typically succeed
on retrying. If you see other types of errors, retrying is unlikely to help.
}
\examples{
\dontrun{
# Register your Qualtrics credentials if you haven't already
qualtrics_api_credentials(
  api_key = "<YOUR-API-KEY>",
  base_url = "<YOUR-BASE-URL>"
)

# Retrieve a list of all mailing lists
mailinglists <- all_mailinglists()

# Retrieve a single mailing list
mailinglist <- fetch_mailinglist(mailinglists$mailingListId[1])
}

}
