% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_description.R
\name{fetch_qsf}
\alias{fetch_qsf}
\title{Download (and potentially save) a description of a survey in Qualtrics Survey
Format (QSF)}
\usage{
fetch_qsf(
  surveyID,
  formatting = c("none", "pretty", "compact"),
  save = FALSE,
  file = NULL
)
}
\arguments{
\item{surveyID}{String. Unique ID for the survey you want to download.
Returned as "id" by the \link{all_surveys} function.}

\item{formatting}{String. Adjustments to the downloaded QSF JSON to be either
more human-readable or more compact.  Options are "none" (default, no
adjustment), "pretty" (more human-readable, using
\code{\link[jsonlite]{prettify}}), or "compact" (more compact, using
\code{\link[jsonlite]{minify}})}

\item{save}{Boolean. Should downloaded QSF be output to R as a json object
(\code{FALSE}, default), or to a file (\code{TRUE})? Even when TRUE,
function still invisibly returns results.}

\item{file}{String. File to write to (including path if needed).  Typically,
would end in .qsf, though not required.  Ignored if \code{save == FALSE}.}
}
\value{
A JSON object equivalent of a Qualtrics QSF.
}
\description{
Download (and potentially save) a description of a survey in Qualtrics Survey
Format (QSF)
}
\details{
If the request to the Qualtrics API made by this function fails, the request
will be retried. If you see these failures on a 500 error (such as a 504
error) be patient while the request is retried; it will typically succeed
on retrying. If you see other types of errors, retrying is unlikely to help.
}
