% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey_options.R
\name{fetch_survey_metadata}
\alias{fetch_survey_metadata}
\title{Get metadata content of survey}
\usage{
fetch_survey_metadata(surveyID)
}
\arguments{
\item{surveyID}{String. Unique ID for the survey whose metadata you wish to
update.}
}
\value{
Data frame listing survey metadata
}
\description{
Get metadata content of survey
}
