% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{generate_url}
\alias{generate_url}
\title{Generate URL for specific API query by type and (if appropriate) ID}
\usage{
generate_url(query, ...)
}
\arguments{
\item{query}{string.  The specific API query desired.  Generally named the
same as associated functions but without underscores, so the request for
\code{fetch_survey()} would be be "fetchsurvey".}

\item{...}{Named elements of URL for specific query desired, such as
\code{surveyID} or \code{mailinglistID}}
}
\value{
Endpoint URL to be passed to querying tools
}
\description{
Generate URL for specific API query by type and (if appropriate) ID
}
\keyword{internal}
