% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getOrganization.R
\name{getOrganization}
\alias{getOrganization}
\title{Retrieve a data frame of general information about an organization}
\usage{
getOrganization(organizationId)
}
\arguments{
\item{organizationId}{This is the Qualtrics assigned Organization Id available from the Qualtrics IDs console.}
}
\value{
Returns a dataframe with general information about an organization.
}
\description{
Retrieve a data frame of general information about an organization
}
\examples{
\dontrun{
# Register your Qualtrics credentials if you haven't already
registerOptions(api_token="<YOUR-API-TOKEN>", root_url="<YOUR-ROOT-URL>")
# Retrieve a data frame with organizational details
org <- getOrganization("organizationId")}
}
\seealso{
See \url{https://api.qualtrics.com/docs} for documentation on the Qualtrics API.
}
