% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSurveyIDs.R
\name{getSurveyIDs}
\alias{getSurveyIDs}
\title{Retrieve a list of all active surveys that you own on qualtrics}
\usage{
getSurveyIDs(headers,
  survey_baseurl = "https://yourdatacenterid.qualtrics.com/API/v3/surveys")
}
\arguments{
\item{headers}{'headers' object - returned by the 'constructHeader' function. See \link[qualtRics]{constructHeader}.}

\item{survey_baseurl}{Base url for your institution (see \url{https://api.qualtrics.com/docs/root-url}. If you do not fill in anything, the function will use the default url. Using your institution-specific url can significantly speed up queries.)}
}
\description{
Retrieve a list of all active surveys that you own on qualtrics
}
\author{
Jasper Ginn
}
\seealso{
See \url{https://api.qualtrics.com/docs} for documentation on the Qualtrics API.
}

