% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSurveyQuestions.R
\name{getSurveyQuestions}
\alias{getSurveyQuestions}
\title{Retrieve a Data Frame Containing Question IDs and Labels}
\usage{
getSurveyQuestions(surveyID)
}
\arguments{
\item{surveyID}{String. Unique ID for the survey you want to download. Returned as 'id' by the \link[qualtRics]{getSurveys} function.}
}
\description{
Retrieve a Data Frame Containing Question IDs and Labels
}
\examples{
\dontrun{
# Register your Qualtrics credentials if you haven't already
qualtrics_api_credentials(api_key = "<YOUR-API-KEY>",
                          base_url = "<YOUR-BASE-URL>")

# Retrieve a list of surveys
surveys <- all_surveys()

# Retrieve questions for a survey
questions <- getSurveyQuestions(surveyID = surveys$id[6])
# Retrieve a single survey, filtering for questions I want.
mysurvey <- getSurvey(surveyID = surveys$id[6],
                      saveDir = tempdir(),
                      includedQuestionIds = c("QID1", "QID2", "QID3"),
                      verbose = TRUE)
}
}
\seealso{
See \url{https://api.qualtrics.com/docs} for documentation on the Qualtrics API.
}
