% Generated by roxygen2 (4.0.1): do not edit by hand
\name{getSurveyResults}
\alias{getSurveyResults}
\title{Returns the results of a given survey.}
\usage{
getSurveyResults(reqAuth, surveyid)
}
\arguments{
\item{reqAuth}{list object returned from \code{\link{auth}}}

\item{surveyid}{the Qualtrics survey id.}
}
\description{
Returns the results of a given survey.
}

