% Generated by roxygen2 (4.0.1): do not edit by hand
\name{importTemplate}
\alias{importTemplate}
\title{import Excel template}
\usage{
importTemplate(XLSpath)
}
\arguments{
\item{XLSpath}{path to your "path/template.xls" file}
}
\description{
This function imports two sheets from the provided Excel template:
"survey" and "choices" and returns a list.
}

