% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{infer_data_types}
\alias{infer_data_types}
\title{Set proper data types on survey data.}
\usage{
infer_data_types(data, surveyID, verbose = FALSE)
}
\arguments{
\item{data}{Imported Qualtrics survey}

\item{surveyID}{ID of survey}

\item{verbose}{Flag}
}
\description{
Set proper data types on survey data.
}
\keyword{internal}
