% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/key_from_file.R
\name{key_from_file}
\alias{key_from_file}
\title{Set the environment variable for the API key from a file}
\usage{
key_from_file(path = path.expand("~/.qualtrics_key"), verbose = FALSE)
}
\arguments{
\item{path}{A file path.}

\item{verbose}{Whether to print the result of reading from \code{path}, or
not (the default).}
}
\value{
Invisibly returns the value of the environment variable
  \code{QUALTRICS_KEY}, after setting it as a side effect.
}
\description{
The environment variable \code{QUALTRICS_KEY} should give an API key for
communication with the Qualtrics survey platform. \code{key_from_file} sets
it using the contents of a file whose first line contains the API key.
}
\details{
The environment variable for the API key can be set during
\link[=Startup]{startup} (recommended), interactively with
\code{\link[base]{Sys.setenv}} or through \code{key_from_file}.
}
