% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_distribution_links.R
\name{list_distribution_links}
\alias{list_distribution_links}
\title{Download distribution links for a distribution from Qualtrics}
\usage{
list_distribution_links(distributionID, surveyID)
}
\arguments{
\item{distributionID}{String. Unique distribution ID for the distribution links you want to download.}
}
\description{
Download distribution links for a distribution from Qualtrics
}
\examples{
\dontrun{
# Register your Qualtrics credentials if you haven't already
qualtrics_api_credentials(
  api_key = "<YOUR-API-KEY>",
  base_url = "<YOUR-BASE-URL>"
)

surveys <- all_surveys()
distributions <- fetch_distributions(surveys$id[1])
distribution_links <- list_distribution_links(distributions$id[1], surveyID = surveys$id[1])
}

}
