% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mailing_list_contacts.R
\name{mailing_list_contacts}
\alias{mailing_list_contacts}
\title{Retrieve a data frame of all contacts for a mailing list from Qualtrics}
\usage{
mailing_list_contacts(directoryID, mailinglistID)
}
\arguments{
\item{directoryID}{String. Unique directory ID.}

\item{mailinglistID}{String. Unique ID of mailing list you want contacts from.}
}
\description{
Retrieve a data frame of all contacts for a mailing list from Qualtrics
}
\details{
If the request to the Qualtrics API made by this function fails, the request
will be retried. If you see these failures on a 500 error (such as a 504
error) be patient while the request is retried; it will typically succeed
on retrying. If you see other types of errors, retrying is unlikely to help.
}
\examples{
\dontrun{
# Register your Qualtrics credentials if you haven't already
qualtrics_api_credentials(
  api_key = "<YOUR-API-KEY>",
  base_url = "<YOUR-BASE-URL>"
)

directories <- list_directories()
mailing_lists <- list_mailing_lists(directories$directoryId[1])
contacts <- mailing_list_contacts(mailing_lists$mailingListId[1])
}

}
