% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{make_colmap}
\alias{make_colmap}
\title{Extracting column mappings from raw downloaded data
helps \code{read_survey} to extract column map attributes to be put w/in response data:}
\usage{
make_colmap(file_name, qnames = NULL, as_dataframe = FALSE)
}
\arguments{
\item{file_name}{file name passed from read_survey}

\item{qnames}{passed in from \code{read_survey}, else gets from file}

\item{as_dataframe}{does output come as a tibble of column mappings (much like column_map currently) or
instead as a list (used internally within \code{read_survey})?}
}
\description{
Extracting column mappings from raw downloaded data
helps \code{read_survey} to extract column map attributes to be put w/in response data:
}
