% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseXMLResponse.r
\name{parseXMLResponse}
\alias{parseXMLResponse}
\title{Internal method. Parses XML response from Qualtrics.}
\usage{
parseXMLResponse(doc)
}
\description{
For Qualtrics functions dates startDate and endDate format: YYYY-MM-DD
}

