% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{qualtrics-package}
\alias{qualtrics-package}
\alias{qualtrics}
\title{The qualtrics package: overview}
\description{
qualtrics is an R client for the Qualtrics survey platform. You will need a
Qualtrics account and an
\href{https://www.qualtrics.com/support/integrations/api-integration/api-integration}{API key}
for most functionality.
}
\author{
Seth Barry

James Dunham <james@jdunham.io>

Jasper Ginn <jasperginn@hotmail.com>
}
\keyword{package}
