% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{qualtrics_api_request}
\alias{qualtrics_api_request}
\title{Send httr requests to Qualtrics API}
\usage{
qualtrics_api_request(verb = c("GET", "POST"), url = url,
  body = NULL)
}
\arguments{
\item{verb}{Type of request to be sent (@seealso \code{\link[httr]{VERB}})}

\item{url}{Qualtrics endpoint URL created by \code{\link{append_root_url}} function}

\item{body}{Options created by \code{\link{create_raw_payload}} function}
}
\description{
Send httr requests to Qualtrics API
}
