% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualtrics_authenticate.R
\name{qualtrics_authenticate}
\alias{qualtrics_authenticate}
\title{Authenticate your api token, data center and oauth credentials}
\usage{
qualtrics_authenticate(type = c("token", "oauth"), use_keychain = FALSE,
  ...)
}
\arguments{
\item{type}{either one of 'token' (API credentials) or 'oauth' (using secret and id)}

\item{use_keychain}{should qualtRics retrieve stored credentials from the keychain?}

\item{...}{any of the options (api_token, data_center, client_id, client_secret) can be passed as named objects. If you pass these parameters, they will override e.g. environment variables.}
}
\value{
exits silently
}
\description{
Authenticate your api token, data center and oauth credentials
}
