% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{qualtrics_response_codes}
\alias{qualtrics_response_codes}
\title{Checks responses against Qualtrics response codes and returns error message.}
\usage{
qualtrics_response_codes(res)
}
\arguments{
\item{res}{results object from httr#' @keywords internal}

\item{raw}{If TRUE, add 'raw' flag to httr::content() function.}
}
\description{
Checks responses against Qualtrics response codes and returns error message.
}
