% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{qualtricsApiRequest}
\alias{qualtricsApiRequest}
\title{Send httr requests to Qualtrics API}
\usage{
qualtricsApiRequest(verb = c("GET", "POST"), url = url, body = NULL)
}
\arguments{
\item{verb}{Type of request to be sent (@seealso \code{\link[httr]{VERB}})}

\item{url}{Qualtrics endpoint URL created by \code{\link{append_root_url}} function}

\item{body}{Options created by \code{\link{createRawPayload}} function}
}
\description{
Send httr requests to Qualtrics API
}
