% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{qualtRicsConfigFile}
\alias{qualtRicsConfigFile}
\title{Prints an Example of a QualtRics Configuration File to the Console.}
\usage{
qualtRicsConfigFile(...)
}
\arguments{
\item{...}{All arguments for \code{qualtRicsConfigFile}}
}
\description{
This function is deprecated; use \code{\link[qualtRics]{qualtrics_api_credentials}} instead.
}
\examples{
\dontrun{
# Execute this line to get instructions on how to make a .qualtrics.yml config file.
qualtRicsConfigFile()
}

}
