% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config_from_file.R
\name{read_config}
\alias{read_config}
\title{Set API environment variables from a configuration file}
\usage{
read_config(path = path.expand("~/.qualtrics_api"), select = NULL,
  verbose = FALSE)
}
\arguments{
\item{path}{A file path.}

\item{select}{Keys to select from the configuration file, optionally.}

\item{verbose}{Whether to output the names of the environment variables read
from the configuration file, or not (the default).}
}
\value{
Invisibly returns the names of the environment variables
  read from the configuration file, after setting them as a side effect.
}
\description{
The environment variable \code{QUALTRICS_TOKEN} should give an
\href{https://api.qualtrics.com/docs/authentication}{API token} for
authentication. \code{QUALTRICS_SUBDOMAIN} should give the subdomain(s) that
precede \code{qualtrics.com} in your
\href{https://api.qualtrics.com/docs/root-url}{API URL}.
\code{read_config} sets these environment variables using the contents of a
JSON-formatted configuration file.
}
\details{
An example configuration file:

\code{
\{
  "QUALTRICS_TOKEN": "secret_token",
  "QUALTRICS_SUBDOMAIN": "co1"
\}
}

Environment variables can be set during \link[=Startup]{startup}
(recommended), interactively with \code{\link[base]{Sys.setenv}} or through
\code{read_config}.
}
