% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readSurvey.R
\name{readSurvey}
\alias{readSurvey}
\title{Read comma separated csv file from Qualtrics.}
\usage{
readSurvey(file_name, stripHTML = TRUE, legacyFormat = FALSE)
}
\arguments{
\item{file_name}{String. A csv data file.}

\item{stripHTML}{Logical. If TRUE, then remove html tags. Defaults to TRUE}

\item{legacyFormat}{Logical. If TRUE, then import "legacy" format csv files (as of 2017). Defaults to FALSE}
}
\value{
A data frame. Variable labels are stored as attributes. They are not printed on
the console but are visibile in the RStudio viewer.
}
\description{
\code{readSurvey}
Reads comma separated csvdata files generated by Qualtrics
software. The second line containing the variable labels is imported.
Repetitive introductions to matrix questions are automatically removed.
Variable labels are stored as attributes.
}
\examples{
\dontrun{
my_data_frame <- readSurvey("<YOUR-PATH-TO-CSV-FILE>")
}
}
\author{
Adrian Brugger, Stefan Borer & Jasper Ginn
}
