% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/registerOptions.R
\name{registerOptions}
\alias{registerOptions}
\title{Register qualtrics API key, root url and other options}
\usage{
registerOptions(api_token = NULL, root_url = NULL)
}
\arguments{
\item{api_token}{String. API token. Available in your qualtrics account (see: \url{https://api.qualtrics.com/docs/authentication})}

\item{root_url}{String. Root url for your institution (see: \url{https://api.qualtrics.com/docs/root-url})}
}
\description{
This function registers the user's qualtrics API key, root url and other options for the remainder of the R session. This function only needs to be called once (at the beginning of each R session).
}
\examples{
\dontrun{
registerApiKey(api_token = "<YOUR-API-TOKEN>", root_url = "<YOUR-INSTITUTION-ROOT-URL>",
               useLocalTimezone = TRUE)
# Retrieve a list of surveys
surveys <- getSurveys(root_url = "https://leidenuniv.eu.qualtrics.com")
                      # URL is for my own institution.
                      # Substitute with your own institution's url
# Retrieve a single survey
mysurvey <- getSurvey(surveyID = surveys$id[6],
                      root_url = "https://leidenuniv.eu.qualtrics.com",
                      save_dir = tempdir(),
                      verbose=TRUE)
# You can use the same parameters as those found in the qualtrics API documentation
# Found here: https://api.qualtrics.com/docs/csv
mysurvey <- getSurvey(surveyID = surveys$id[6],
                      root_url = "https://leidenuniv.eu.qualtrics.com",
                      save_dir = tempdir(),
                      startDate = "2017-01-01",
                      endDate = "2017-01-31",
                      limit = 100,
                      useLabels = TRUE,
                      seenUnansweredRecode = "UNANS",
                      verbose=TRUE)
}

}
\seealso{
See \url{https://github.com/JasperHG90/qualtRics/blob/master/README.md#using-a-configuration-file} for more information about the qualtRics configuration file.
}
\author{
Jasper Ginn
}
