% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/request.R
\name{request}
\alias{request}
\title{Send a request to the Qualtrics API}
\usage{
request(verb = c("GET", "POST", "PUT", "DELETE"), action,
  key = Sys.getenv("QUALTRICS_KEY"),
  subdomain = Sys.getenv("QUALTRICS_SUBDOMAIN"), verbose = FALSE, ...)
}
\arguments{
\item{verb}{Name of verb to use.}

\item{action}{Either an API action (like \code{"surveys"}) or a complete API
URL.}

\item{key}{A Qualtrics API key (by default, the value of the environment
variable \code{QUALTRICS_KEY}).}

\item{subdomain}{A Qualtrics subdomain (by default, the value of the
environment variable \code{QUALTRICS_SUBDOMAIN}). Ignored if `action` is a
complete API URL.}

\item{verbose}{Output API calls to \code{\link[base]{stderr}}.}

\item{...}{Further arguments to the function given by \code{verb}, e.g.,
\code{\link[httr]{GET}}.}
}
\value{
For \code{request}, a \code{\link[httr]{response}} object. For
  \code{qget} and \code{qpost}, its content as extracted by
  \code{\link[httr]{content}}.
}
\description{
`request` is a lower-level function for direct access to API actions. It may
be useful in programming.
}
\details{
For help with subdomains, see the
\href{https://api.qualtrics.com/docs/root-url}{Qualtrics documentation}. Each
Qualtrics account is assigned a subdomain, and using another will work but
produce a warning.
}
\examples{
\dontrun{
response <- request("GET", "surveys")
}
}

