% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey_options.R
\name{update_survey_metadata}
\alias{update_survey_metadata}
\title{Update metadata options}
\usage{
update_survey_metadata(
  surveyID,
  name = NULL,
  description = NULL,
  active = NULL,
  active_start_date = NULL,
  active_end_date = NULL,
  time_zone = NULL
)
}
\arguments{
\item{surveyID}{String. Unique ID for the survey whose metadata you wish to
update.}

\item{name}{String. Set survey name.  Defaults to NULL, which leaves existing
name.}

\item{description}{String. Set survey description. Defaults to \code{NULL}, which
leaves existing description.  Entering \code{NA} will remove any existing
description.}

\item{active}{Boolean. Set survey activity status (whether open for
collection).  Defaults to \code{NULL}, which leaves status unchanged.}

\item{active_start_date}{POSIXlt, POSIXct, Date, or string equivalent.  Sets
date/time for survey becoming active.  See Details.}

\item{active_end_date}{POSIXlt, POSIXct, Date, or string equivalent.  Sets
date/time for survey becoming inactive  See Details.}

\item{time_zone}{String. Desired time zone to use when \code{active_start_date} or
\code{active_end_date} are given as Date or string objects.  Ignored if argument(s)
given as POSIXlt/POSIXct objects. Must match a time zone name from
\code{\link[base:timezones]{base::OlsonNames()}}. Defaults to \code{NULL}, which uses the current
system timezone (\code{\link[base:timezones]{base::Sys.timezone()}}).}
}
\description{
Update metadata options
}
