% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_description.R
\name{write_qsf}
\alias{write_qsf}
\title{Download and save a description of a survey in Qualtrics Survey Format (QSF)}
\usage{
write_qsf(
  surveyID,
  formatting = c("none", "pretty", "compact"),
  file = NULL,
  save = TRUE
)
}
\arguments{
\item{surveyID}{String. Unique ID for the survey you want to download.
Returned as "id" by the \link{all_surveys} function.}

\item{formatting}{String. Adjustments to the downloaded QSF JSON to be either
more human-readable or more compact.  Options are "none" (default, no
adjustment), "pretty" (more human-readable, using
\code{\link[jsonlite]{prettify}}), or "compact" (more compact, using
\code{\link[jsonlite]{minify}})}

\item{file}{String. Desired filename, including path if desired.
Typical filename extension is .qsf, though not required (and will not be
added if absent).  Argument ignored if \code{save == FALSE}.}

\item{save}{Boolean. Should downloaded QSF be saved to a file (\code{TRUE},
default), or returned as output (\code{FALSE})?  Even when \code{TRUE},
downloaded JSON is returned invisibly.}
}
\value{
A JSON object equivalent of a Qualtrics QSF, invisibly unless
\code{save == FALSE}
}
\description{
Download and save a description of a survey in Qualtrics Survey Format (QSF)
}
\details{
If the request to the Qualtrics API made by this function fails, the request
will be retried. If you see these failures on a 500 error (such as a 504
error) be patient while the request is retried; it will typically succeed
on retrying. If you see other types of errors, retrying is unlikely to help.
}
