% Generated by roxygen2 (4.0.1): do not edit by hand
\name{writeQtxt}
\alias{writeQtxt}
\title{create an advanced format .txt survey file to import into Qualtrics}
\usage{
writeQtxt(Qlist, f = "qualtrics.txt")
}
\arguments{
\item{Qlist}{the object created by \code{\link{importTemplate}}}

\item{f}{the path and name of the .txt file you want to generate; default is "qualtrics.txt"}
}
\description{
This function creates a .txt file formatted for direct import into
Qualtrics \url{http://www.qualtrics.com/university/researchsuite/advanced-building/advanced-options-drop-down/import-and-export-surveys/}.
}

