% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{construct_header}
\alias{construct_header}
\title{Construct a header to send to Qualtrics API}
\usage{
construct_header(API_TOKEN)
}
\arguments{
\item{API_TOKEN}{API token. Available in your Qualtrics account (see: \url{https://api.qualtrics.com/})}
}
\description{
Construct a header to send to Qualtrics API
}
