% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_raw_payload}
\alias{create_raw_payload}
\title{Create raw JSON payload to post response exports request}
\usage{
create_raw_payload(surveyID, label = TRUE, last_response = NULL,
  start_date = NULL, end_date = NULL, limit = NULL,
  local_time = FALSE, unanswer_recode = NULL,
  include_questions = NULL)
}
\arguments{
\item{surveyID}{Survey ID}

\item{label}{Flag}

\item{last_response}{Flag}

\item{start_date}{Flag}

\item{end_date}{Flag}

\item{limit}{Flag}

\item{local_time}{Flag}

\item{unanswer_recode}{Flag}

\item{include_questions}{Flag}
}
\value{
JSON file with options to send to API
}
\description{
Create raw JSON payload to post response exports request
}
\seealso{
Functions \code{\link{all_surveys}} and \code{\link{registerOptions}}
have more details on these parameters
}
