% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_raw_payload}
\alias{create_raw_payload}
\title{Create raw JSON payload to post response exports request}
\usage{
create_raw_payload(
  label = TRUE,
  start_date = NULL,
  end_date = NULL,
  limit = NULL,
  time_zone = NULL,
  unanswer_recode = NULL,
  unanswer_recode_multi = NULL,
  include_displayorder = TRUE,
  newline_string = NULL,
  include_questions = NULL
)
}
\arguments{
\item{label}{Flag}

\item{start_date}{Flag}

\item{end_date}{Flag}

\item{limit}{Flag}

\item{unanswer_recode}{Flag}

\item{include_questions}{Flag}

\item{surveyID}{Survey ID}

\item{last_response}{Flag}

\item{local_time}{Flag}
}
\value{
JSON file with options to send to API
}
\description{
Create raw JSON payload to post response exports request
}
\seealso{
Functions \code{\link{all_surveys}} and \code{\link{registerOptions}}
have more details on these parameters
}
