% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_raw_payload}
\alias{create_raw_payload}
\title{Create properly-formatted JSON payload for API calls.  Removes NULLS}
\usage{
create_raw_payload(...)
}
\value{
JSON file with options to send to API
}
\description{
Create properly-formatted JSON payload for API calls.  Removes NULLS
}
\keyword{internal}
