% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_raw_payload}
\alias{create_raw_payload}
\title{Create raw JSON payload to post response exports request}
\usage{
create_raw_payload(surveyID, useLabels = TRUE, lastResponseId = NULL,
  startDate = NULL, endDate = NULL, limit = NULL,
  useLocalTime = FALSE, seenUnansweredRecode = NULL,
  includedQuestionIds = NULL)
}
\arguments{
\item{surveyID}{Survey ID}

\item{useLabels}{Flag}

\item{lastResponseId}{Flag}

\item{startDate}{Flag}

\item{endDate}{Flag}

\item{limit}{Flag}

\item{useLocalTime}{Flag}

\item{seenUnansweredRecode}{Flag}

\item{includedQuestionIds}{Flag}
}
\value{
JSON file with options to send to API
}
\description{
Create raw JSON payload to post response exports request
}
\seealso{
Functions \code{\link{getSurveys}} and \code{\link{registerOptions}}
have more details on these parameters
}
