% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{download_qualtrics_export}
\alias{download_qualtrics_export}
\title{Download response export}
\usage{
download_qualtrics_export(fetch_url, requestID, verbose = FALSE)
}
\arguments{
\item{fetch_url}{URL provided by Qualtrics API that shows the download percentage completeness}

\item{requestID}{ID}

\item{verbose}{See \code{\link{fetch_survey}}}
}
\description{
Download response export
}
\details{
If the request to the Qualtrics API made by this function fails, the request
will be retried. If you see these failures on a 500 error (such as a 504
error) be patient while the request is retried; it will typically succeed
on retrying.
}
\keyword{internal}
