% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{export_responses_filedownload}
\alias{export_responses_filedownload}
\title{Downloads response data (.zip of .csv) from location obtained from
fetch_survey_progress}
\usage{
export_responses_filedownload(surveyID, fileID)
}
\arguments{
\item{surveyID}{survey ID}

\item{fileID}{file ID from fetch_survey_progress}

\item{requestID}{request ID from fetch_survey}
}
\description{
Downloads response data (.zip of .csv) from location obtained from
fetch_survey_progress
}
\keyword{internal}
