% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_colmap.R
\name{extract_colmap}
\alias{extract_colmap}
\title{Test function: a way of extracting column mappings from existing dataframes}
\usage{
extract_colmap(r_dat)
}
\arguments{
\item{r_dat}{response data obtained from fetch_survey w/\code{colmap_attrs = TRUE}}
}
\description{
could be used as both helper function and as a replacement for column_map, once
updated to also generate column mappings from existing response download data frames,
or from saved response downloads, etc.
}
