% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_distribution_history.R
\name{fetch_distribution_history}
\alias{fetch_distribution_history}
\title{Download distribution history data for a distribution from Qualtrics}
\usage{
fetch_distribution_history(distributionID, surveyID)
}
\arguments{
\item{distributionID}{String. Unique distribution ID for the distribution history you want to download.}

\item{surveyID}{String. Unique ID for the survey you want to download.
Returned as \code{id} by the \link[qualtRics]{all_surveys} function.}
}
\description{
Download distribution history data for a distribution from Qualtrics
}
\examples{
\dontrun{
# Register your Qualtrics credentials if you haven't already
qualtrics_api_credentials(
  api_key = "<YOUR-API-KEY>",
  base_url = "<YOUR-BASE-URL>"
)

surveys <- all_surveys()
distributions <- fetch_distributions(surveys$id[1])
distribution_history <- fetch_distribution_history(
    distributions$id[1],
    surveyID = surveys$id[1]
)
}

}
