% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_mailinglist.R
\name{fetch_mailinglist}
\alias{fetch_mailinglist}
\title{Download a mailing list from Qualtrics}
\usage{
fetch_mailinglist(mailinglistID)
}
\arguments{
\item{mailinglistID}{String. Unique ID for the mailing list you want to
download. Returned as \code{id} by the \link[=all_mailinglists]{all_mailinglists}
function.}
}
\description{
Download a mailing list from Qualtrics
}
\details{
If the request to the Qualtrics API made by this function fails, the request
will be retried. If you see these failures on a 500 error (such as a 504
error) be patient while the request is retried; it will typically succeed
on retrying. If you see other types of errors, retrying is unlikely to help.
}
\examples{
\dontrun{
# Register your Qualtrics credentials if you haven't already
qualtrics_api_credentials(
  api_key = "<YOUR-API-KEY>",
  base_url = "<YOUR-BASE-URL>"
)

# Retrieve a list of all mailing lists
mailinglists <- all_mailinglists()

# Retrieve a single mailing list
mailinglist <- fetch_mailinglist(mailinglists$id[1])
}

}
