% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panels.R
\name{getPanel}
\alias{getPanel}
\title{Returns the survey invitees for the given panel.}
\usage{
getPanel(username, password, libraryid, panelid, embeddedData = NULL)
}
\arguments{
\item{username}{the Qualtrics username.}

\item{password}{the Qualtrics password.}

\item{libraryid}{the Qualtrics library id.}

\item{panelid}{the Qualtrics panel id.}

\item{embeddedData}{embedded data to return.}
}
\description{
Returns the survey invitees for the given panel.
}

