% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panels.R
\name{getPanelMemberCount}
\alias{getPanelMemberCount}
\title{Returns the number of members in the panel.}
\usage{
getPanelMemberCount(username, password, libraryid, panelid)
}
\arguments{
\item{username}{the Qualtrics username.}

\item{password}{the Qualtrics password.}

\item{libraryid}{the Qualtrics library id.}

\item{panelid}{the Qualtrics panel id.}
}
\description{
Returns the number of members in the panel.
}

