% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panels.R
\name{getPanels}
\alias{getPanels}
\title{Returns the survey panels in the given library.}
\usage{
getPanels(qualtrics.token, qualtrics.baseurl = "https://qualtrics.com",
  verbose = FALSE)
}
\arguments{
\item{qualtrics.token}{authentication token.}

\item{qualtrics.baseurl}{The base URL for the Qualtrics API. See 
https://api.qualtrics.com/docs/root-url for more information about
how to determine your base URL.}

\item{verbose}{if TRUE, a statement will be printed when retrieving 
subsequent pages.}
}
\description{
Returns the survey panels in the given library.
}

