% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panels.R
\name{getPanels}
\alias{getPanels}
\title{Returns the survey panels in the given library.}
\usage{
getPanels(username, password, libraryid)
}
\arguments{
\item{username}{the Qualtrics username.}

\item{password}{the Qualtrics password.}

\item{libraryid}{the Qualtrics library id.}
}
\description{
Returns the survey panels in the given library.
}

