\name{getRecipient}
\alias{getRecipient}
\title{Returns a recipient from the panel.}
\usage{
  getRecipient(username, password, recipientId, libraryId)
}
\arguments{
  \item{username}{the Qualtrics username.}

  \item{password}{the Qualtrics password.}

  \item{libraryId}{the Qualtrics library id.}

  \item{recipientId}{the id for the recipient to return.}
}
\description{
  Returns a recipient from the panel.
}

