% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panels.R
\name{getRecipient}
\alias{getRecipient}
\title{Returns a recipient from the panel.}
\usage{
getRecipient(username, password, recipientId, libraryId)
}
\arguments{
\item{username}{the Qualtrics username.}

\item{password}{the Qualtrics password.}

\item{recipientId}{the id for the recipient to return.}

\item{libraryId}{the Qualtrics library id.}
}
\description{
Returns a recipient from the panel.
}

