\name{getSurvey}
\alias{getSurvey}
\title{Returns the given survey.}
\usage{
  getSurvey(username, password, surveyid)
}
\arguments{
  \item{username}{the Qualtrics username.}

  \item{password}{the Qualtrics password.}

  \item{surveyid}{the Qualtrics survey id.}
}
\description{
  Returns the given survey.
}

