% Generated by roxygen2 (4.0.1): do not edit by hand
\name{getSurvey}
\alias{getSurvey}
\title{Returns the given survey.}
\usage{
getSurvey(reqAuth, surveyid)
}
\arguments{
\item{reqAuth}{list object returned from \code{\link{auth}}}

\item{surveyid}{the Qualtrics survey id.}
}
\description{
Returns the given survey.
}

