% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_survey.R
\name{getSurvey}
\alias{getSurvey}
\title{Download a survey and import it into R}
\usage{
getSurvey(...)
}
\arguments{
\item{...}{All arguments for \code{fetch_survey}}
}
\description{
This function is soft deprecated; use \code{\link[qualtRics]{fetch_survey}}
instead.
}
