% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSurvey.R
\name{getSurvey}
\alias{getSurvey}
\title{Export a survey and download it into R}
\usage{
getSurvey(surveyID, lastResponseId = NULL, startDate = NULL,
  endDate = NULL, seenUnansweredRecode = NULL, limit = NULL,
  includedQuestionIds = NULL, save_dir = NULL, force_request = FALSE,
  verbose = TRUE, useLabels = TRUE, convertVariables = TRUE,
  useLocalTime = FALSE, ...)
}
\arguments{
\item{surveyID}{String. Unique ID for the survey you want to download. Returned as 'id' by the \link[qualtRics]{getSurveys} function.}

\item{lastResponseId}{String. Export all responses received after the specified response. Defaults to NULL.}

\item{startDate}{String. Filter to only exports responses recorded after the specified date. Accepts dates as character strings in format "YYYY-MM-DD". Defaults to NULL.}

\item{endDate}{String. Filter to only exports responses recorded before the specified date. Accepts dates as character strings in format "YYYY-MM-DD". Defaults to NULL.}

\item{seenUnansweredRecode}{String. Recode seen but unanswered questions with a string value. Defaults to NULL.}

\item{limit}{Integer. Maximum number of responses exported. Defaults to NULL (all responses).}

\item{includedQuestionIds}{Vector of strings (e.g. c('QID1', 'QID2', 'QID3'). Export only specified questions. Defaults to NULL.}

\item{save_dir}{String. Directory where survey results will be stored. Defaults to a temporary directory which is cleaned when your R session is terminated. This argument is useful if you'd like to store survey results. The downloaded survey will be stored as an RDS file (see \link[base]{readRDS}).}

\item{force_request}{Logical. getSurvey() saves each survey in a temporary directory so that it can quickly be retrieved later. If force_request is TRUE, getSurvey() always downloads the survey from the API instead of loading it from the temporary directory. Defaults to FALSE.}

\item{verbose}{Logical. If TRUE, verbose messages will be printed to the R console. Defaults to TRUE.}

\item{useLabels}{Logical. TRUE to export survey responses as Choice Text or FALSE to export survey responses as values.}

\item{convertVariables}{Logical. If TRUE, then the \code{\link[qualtRics]{getSurvey}} function will convert certain question types (e.g. multiple choice) to proper data type in R. Defaults to TRUE.}

\item{useLocalTime}{Logical. Use local timezone to determine response date values? Defaults to FALSE. See \url{https://api.qualtrics.com/docs/dates-and-times} for more information.}

\item{...}{optional arguments. You can pass all arguments listed in \code{\link{registerOptions}} (except a different base url / api key). You can also pass a argument 'fileEncoding' (see 'fileEncoding' argument in \code{\link{readSurvey}}) to import your survey using a specific encoding.}
}
\description{
Export a Qualtrics survey you own and import the survey directly into R.
}
\examples{
\dontrun{
# Register your Qualtrics credentials if you haven't already
qualtrics_api_credentials(api_key = "<YOUR-API-KEY>",
                          base_url = "<YOUR-BASE-URL>")

# Retrieve a list of surveys
surveys <- getSurveys()

# Retrieve a single survey
mysurvey <- getSurvey(surveyID = surveys$id[6])

# You can use the same parameters as those found in the Qualtrics API documentation
# Found here: https://api.qualtrics.com/docs/csv
mysurvey <- getSurvey(surveyID = surveys$id[6],
                      save_dir = tempdir(),
                      startDate = "2018-01-01",
                      endDate = "2018-01-31",
                      limit = 100,
                      useLabels = TRUE,
                      seenUnansweredRecode = "UNANS",
                      verbose = TRUE)
}
}
\seealso{
See \url{https://api.qualtrics.com/docs/csv} for documentation on the Qualtrics API.
}
