% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surveys.R
\name{getSurveyName}
\alias{getSurveyName}
\title{Returns the given survey name.}
\usage{
getSurveyName(username, password, surveyid)
}
\arguments{
\item{username}{the Qualtrics username.}

\item{password}{the Qualtrics password.}

\item{surveyid}{the Qualtrics survey id.}
}
\description{
Returns the given survey name.
}

