% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{getSurveyQuestions}
\alias{getSurveyQuestions}
\title{Retrieve a data frame containing question IDs and labels}
\usage{
getSurveyQuestions(...)
}
\arguments{
\item{...}{All arguments for \code{survey_questions}}
}
\description{
This function is deprecated; use \code{\link[qualtRics]{survey_questions}}
instead.
}
