% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey_questions.R
\name{getSurveyQuestions}
\alias{getSurveyQuestions}
\title{Retrieve a data frame containing question IDs and labels}
\usage{
getSurveyQuestions(surveyID)
}
\arguments{
\item{surveyID}{A string. Unique ID for the survey you want to download.
Returned as `id` by the \link[qualtRics]{getSurveys} function.}
}
\description{
This function is soft deprecated; use \code{\link[qualtRics]{survey_questions}}
instead.
}
