% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSurveyQuestions.R
\name{getSurveyQuestions}
\alias{getSurveyQuestions}
\title{Retrieve a data frame containing question IDs and labels}
\usage{
getSurveyQuestions(surveyID,
  root_url = "https://yourdatacenterid.qualtrics.com")
}
\arguments{
\item{surveyID}{String. Unique ID for the survey you want to download. Returned as 'id' by the \link[qualtRics]{getSurveys} function.}

\item{root_url}{String. Base url for your institution (see \url{https://api.qualtrics.com/docs/csv}. You need to supply this url. Your query will NOT work without it.).}
}
\description{
Retrieve a data frame containing question IDs and labels
}
\examples{
\dontrun{
registerApiKey("<YOUR-QUALTRICS-API-KEY>")
# Retrieve a list of surveys
surveys <- getSurveys(root_url = "https://leidenuniv.eu.qualtrics.com")
                      # URL is for my own institution.
                      # Substitute with your own institution's url
# Retrieve questions for a survey
questions <- getSurveyQuestions(surveyID = surveys$id[6],
                                root_url = "https://leidenuniv.eu.qualtrics.com")
# Retrieve a single survey, filtering for questions I want.
mysurvey <- getSurvey(surveyID = surveys$id[6],
                      root_url = "https://leidenuniv.eu.qualtrics.com",
                      save_dir = tempdir(),
                      includedQuestionIds = c("QID1", "QID2", "QID3"),
                      verbose=TRUE)
}
}
\seealso{
See \url{https://api.qualtrics.com/docs} for documentation on the Qualtrics API.
}
\author{
Jasper Ginn
}
