% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surveys.R
\name{getSurveys}
\alias{getSurveys}
\title{Returns a list of surveys available for the given user.}
\usage{
getSurveys(qualtrics.token, qualtrics.baseurl = "https://qualtrics.com")
}
\description{
Returns a list of surveys available for the given user.
}

