\name{getSurveys}
\alias{getSurveys}
\title{Returns a list of surveys available for the given user.}
\usage{
  getSurveys(username, password)
}
\arguments{
  \item{username}{the Qualtrics username.}

  \item{password}{the Qualtrics password.}
}
\description{
  Returns a list of surveys available for the given user.
}

