% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surveys.R
\name{getSurveys}
\alias{getSurveys}
\title{Returns a list of surveys available for the given user.}
\usage{
getSurveys(username, password)
}
\arguments{
\item{username}{the Qualtrics username.}

\item{password}{the Qualtrics password.}
}
\description{
Returns a list of surveys available for the given user.
}

