% Generated by roxygen2 (4.0.1): do not edit by hand
\name{getSurveys}
\alias{getSurveys}
\title{Returns a list of surveys available for the given user.}
\usage{
getSurveys(reqAuth)
}
\arguments{
\item{reqAuth}{list object returned from \code{\link{auth}}}
}
\description{
Returns a list of surveys available for the given user.
}

