% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{getSurveys}
\alias{getSurveys}
\title{Retrieve a data frame of all active surveys on Qualtrics}
\usage{
getSurveys()
}
\description{
This function is deprecated; use \code{\link[qualtRics]{all_surveys}}
instead.
}
