% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_distribution_links.R
\name{list_distribution_links}
\alias{list_distribution_links}
\title{Download distribution links for a distribution from Qualtrics}
\usage{
list_distribution_links(distributionID, surveyID)
}
\arguments{
\item{distributionID}{String. Unique distribution ID for the distribution links you want to download.}

\item{surveyID}{String. Unique ID for the survey you want to download.}
}
\description{
Download distribution links for a distribution from Qualtrics
}
\details{
If the request to the Qualtrics API made by this function fails, the request
will be retried. If you see these failures on a 500 error (such as a 504
error) be patient while the request is retried; it will typically succeed
on retrying. If you see other types of errors, retrying is unlikely to help.
}
\examples{
\dontrun{
# Register your Qualtrics credentials if you haven't already
qualtrics_api_credentials(
  api_key = "<YOUR-API-KEY>",
  base_url = "<YOUR-BASE-URL>"
)

surveys <- all_surveys()
distributions <- fetch_distributions("SV_xxxxxxx1")
distribution_links <- list_distribution_links(distributions$id[1], surveyID = "SV_xxxxxxx1")
}

}
