% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{qualtrics-package}
\alias{qualtrics}
\alias{qualtrics-package}
\title{qualtrics}
\description{
An R client for Qualtrics
}
\details{
This is a template package for the cloudyr project
}
\author{
Seth Barry

James Dunham <james@jdunham.io>

Jasper Ginn <jasperginn@hotmail.com>
}
\keyword{package}

