% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualtRicsConfigFile.R
\name{qualtRicsConfigFile}
\alias{qualtRicsConfigFile}
\title{Prints an example of a qualtRics configuration file to the console.}
\usage{
qualtRicsConfigFile(api_token = NULL, root_url = NULL)
}
\description{
Prints an example of a qualtRics configuration file to the console.
}
\examples{
\dontrun{
 qualtRicsConfigFile()
}

}
\seealso{
See \url{https://api.qualtrics.com/docs/root-url} for documentation on the Qualtrics API.
}
\author{
Jasper Ginn
}
