% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualtRicsConfigFile.R
\name{qualtRicsConfigFile}
\alias{qualtRicsConfigFile}
\title{Prints an example of a qualtRics configuration file to the console.}
\usage{
qualtRicsConfigFile(api_token = NULL, root_url = NULL)
}
\arguments{
\item{api_token}{String. API token. Available in your qualtrics account (see: \url{https://api.qualtrics.com/docs/authentication})}

\item{root_url}{String. Root url for your institution (see: \url{https://api.qualtrics.com/docs/root-url})}
}
\description{
Prints an example of a qualtRics configuration file to the console.
}
\examples{
\dontrun{
 qualtRicsConfigFile()
}

}
\seealso{
See \url{https://api.qualtrics.com/docs/root-url} for documentation on the Qualtrics API. See \url{https://github.com/JasperHG90/qualtRics/blob/master/README.md#using-a-configuration-file} for more information about the qualtRics configuration file.
}
\author{
Jasper Ginn
}
