% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_survey.R
\name{readSurvey}
\alias{readSurvey}
\title{Read a CSV file exported from Qualtrics}
\usage{
readSurvey(...)
}
\arguments{
\item{...}{All arguments for \code{\link[qualtRics]{read_survey}}}
}
\description{
Reads comma separated CSV files generated by Qualtrics
software. The second line containing the variable labels is imported.
Repetitive introductions to matrix questions are automatically removed.
Variable labels are stored as attributes.
}
