% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/registerApiKey.R
\name{registerApiKey}
\alias{registerApiKey}
\title{Register qualtrics API key}
\usage{
registerApiKey(API.TOKEN)
}
\arguments{
\item{API.TOKEN}{String. API token. Available in your qualtrics account (see: \url{https://api.qualtrics.com/docs/authentication})}
}
\description{
This function registers the user's qualtrics API key for the remainder of the R session. This function only needs to be called once (at the beginning of each R session).
}
\examples{
\dontrun{
registerApiKey("<YOUR-QUALTRICS-API-KEY>")
# Retrieve a list of surveys
surveys <- getSurveys(root_url = "https://leidenuniv.eu.qualtrics.com")
                      # URL is for my own institution.
                      # Substitute with your own institution's url
# Retrieve a single survey
mysurvey <- getSurvey(surveyID = surveys$id[6],
                      root_url = "https://leidenuniv.eu.qualtrics.com",
                      save_dir = tempdir(),
                      verbose=TRUE)
# You can use the same parameters as those found in the qualtrics API documentation
# Found here: https://api.qualtrics.com/docs/csv
mysurvey <- getSurvey(surveyID = surveys$id[6],
                      root_url = "https://leidenuniv.eu.qualtrics.com",
                      save_dir = tempdir(),
                      startDate = "2017-01-01",
                      endDate = "2017-01-31",
                      limit = 100,
                      useLabels = TRUE,
                      seenUnansweredRecode = "UNANS",
                      verbose=TRUE)
}

}
\seealso{
See \url{https://api.qualtrics.com/docs/root-url} for documentation on the Qualtrics API.
}
\author{
Jasper Ginn
}
