% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{wrapper_mc}
\alias{wrapper_mc}
\title{Convert multiple choice questions to ordered factors}
\usage{
wrapper_mc(data, question_meta)
}
\arguments{
\item{data}{Imported Qualtrics survey}

\item{question_meta}{Question metadata}
}
\description{
Convert multiple choice questions to ordered factors
}
\keyword{internal}
